Fix #78: Calculate estimated vehicle consumption (#79)
* Fix #78: Calculate estimated vehicle consumption

Closes #78

Calculate estimated vehicle consumption. This is pretty fuzzy as it works with charges at home only, depends on how often the car is plugged in as the odometer is only updated when plugging in, and there are some challenges calculating vehicle charge energy.

* Removed vehicle stats from month as monthly values are too fuzzy

* Removed vehicle stats from month

Monthly statistics are way too inaccurate. Removed from month.

Added some informations to yearly stat widgets.